feat: remove changes from migration
