+++ annotation/alembic/versions/71095b8e6343_add_extensive_coverage_parameter.py
@@ -1,0 +1,28 @@
+"""add extensive_coverage parameter
+
+Revision ID: 71095b8e6343
+Revises: 66cd6054c2d0
+Create Date: 2022-12-12 13:36:08.659682
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '71095b8e6343'
+down_revision = '66cd6054c2d0'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('jobs', sa.Column('extensive_coverage', sa.INTEGER(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('jobs', 'extensive_coverage')
+    # ### end Alembic commands ###
