--- annotation/alembic/versions/9befecdd0c9e_add_extensive_coverage_param.py
@@ -1,48 +1,0 @@
-"""add extensive_coverage param
-
-Revision ID: 9befecdd0c9e
-Revises: 66cd6054c2d0
-Create Date: 2022-12-09 13:06:47.578233
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '9befecdd0c9e'
-down_revision = '66cd6054c2d0'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('files', 'status',
-               existing_type=postgresql.ENUM('pending', 'annotated', 'validated', name='file_status'),
-               type_=postgresql.ENUM('pending', 'annotated', 'validated', name='filestatusenumschema'),
-               existing_nullable=False,
-               existing_server_default=sa.text("'pending'::file_status"))
-    op.add_column('jobs', sa.Column('extensive_coverage', sa.INTEGER(), nullable=True))
-    op.alter_column('jobs', 'validation_type',
-               existing_type=postgresql.ENUM('cross', 'hierarchical', 'validation_only', name='validation_type'),
-               type_=postgresql.ENUM('cross', 'hierarchical', 'validation_only', 'extensive_coverage', name='validationschema'),
-               existing_nullable=False,
-               existing_server_default=sa.text("'cross'::validation_type"))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('jobs', 'validation_type',
-               existing_type=postgresql.ENUM('cross', 'hierarchical', 'validation_only', 'extensive_coverage', name='validationschema'),
-               type_=postgresql.ENUM('cross', 'hierarchical', 'validation_only', name='validation_type'),
-               existing_nullable=False,
-               existing_server_default=sa.text("'cross'::validation_type"))
-    op.drop_column('jobs', 'extensive_coverage')
-    op.alter_column('files', 'status',
-               existing_type=postgresql.ENUM('pending', 'annotated', 'validated', name='filestatusenumschema'),
-               type_=postgresql.ENUM('pending', 'annotated', 'validated', name='file_status'),
-               existing_nullable=False,
-               existing_server_default=sa.text("'pending'::file_status"))
-    # ### end Alembic commands ###
