+++ lib/filter_lib/src/__init__.py
@@ -1,4 +1,4 @@
 from .dict_parser import map_request_to_filter  # noqa
 from .pagination import PaginationParams, paginate  # noqa
 from .query_modificator import form_query  # noqa
-from .schema_generator import BaseSearch, Page, create_filter_model   # noqa
+from .schema_generator import BaseSearch, Page, create_filter_model  # noqa
